Huge reductions in string copies.
The main change is that the library is now far less eager to allocate
than it previously was. Allocations will mostly just be for what is used
now.

Context:

`response` and `request` fields have been split into
- `response`, `response_mut`, `request`, and `request_mut` methods.

Request:

These changes were made to make requests "cost less" on requests, a
minimal amount of information is parsed out of the request buffer at the
time of the request, and subsequent calls into `Request` will yield the
post-parse cached value rather than re-parse the values.

Some hash maps have been removed and replaced with vectors of pairs. It
seems while a hash map is more algorithmically ideal, a linear search is
still faster.

The `query` field is now the `query_raw` field.

Add the `header_raw_lines` field.

The following fields have been turned into functions that lazily parse
the values out of the request buffer
- `headers`
- `cookies`
- `get`
- `post`
- `host`
- `user_agent`
- `content_type`
- `content_length`

Utils:

The parser for querystrings and cookie strings has been re-written.

Tests:

Changes to tests to use the updated API. Removed a test for UTF-8
noncompliance in an input, where it is no longer relevant.